Revamp messaging to allow messaging from community page.
Rather than passing a negative thread id to indicate
need for new thread, the server will now always check
if a thread with the requested participants exists.
If one does, it will append the new message to that
thread. If not, it will automatically create a new
thread. This means we no longer have to load all threads
on a profile page for a user, because there's no need to
know up front if a thread already exists if you want
to append a message to it. You just need to know the
participants.