Add 'enact' install script to ensure node_sass module gets installed
Node-sass-middleware has a dependency module, node-sass, which
needs to get installed first. Unfortunately, for some reason,
the current version of node-sass does not properly run
its install script to download the binary file it needs to
operate. See https://github.com/sass/node-sass/issues/1579
for more details.

To fix this, we hadd an 'enact' script to elasticbeanstalk
configuration, to tell it to run the node-sass installer
manually during deployment. This ensures that the binary
file gets installed.