Add authentication for API
Using jsonwebtoken authentication strategy. Client
sends email and password in body request. Server
responds with a token, which can be used on future api
requests. Note this requires us disabling the CSRF
protections on api routes, since the POST will happen
without a CSRF token. This is fine, though, since the api
is not meant to be consumed via a browser.